requirements.txt
amqp==5.1.1
aniso8601==9.0.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.3.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.1.8
django-allauth==0.54.0
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-csp==3.7
django-filter==23.1
django-timezone-field==5.0
djangorestframework==3.14.0
drf-yasg==1.21.5
et-xmlfile==1.1.0
Faker==18.4.0
flake8==6.0.0
future==0.18.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
graphene==3.2.2
graphene-django==3.0.2
graphql-core==3.2.3
graphql-relay==3.2.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
numpy==1.24.2
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.0
promise==2.3
prompt-toolkit==3.0.38
protobuf==4.22.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
rabbitmq==0.2.0
redis==4.5.4
requests==2.28.2
requests-oauthlib==1.3.1
rest-social-auth==8.0.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
serializers==0.2.4
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
text-unidecode==1.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
